# book-network/src/main/java/com/yaroslav/booknetwork/auth/AuthenticationResponse.java
package com.yaroslav.booknetwork.auth;


import lombok.Builder;
import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
@Builder
public class AuthenticationResponse {

    private String token;

}