# book-network/src/main/java/com/yaroslav/booknetwork/feedback/FeedbackResponse.java
package com.yaroslav.booknetwork.feedback;

import lombok.*;

@Getter
@Setter
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class FeedbackResponse {

    private Double note;
    private String comment;
    private boolean ownFeedback;
}